analysis/requirements.txt
apache-flink==1.16.0
kafka-python
websockets
